invoice info node on payment

+++ lib/src/model/PaymentTransaction.dart
@@ -1,7 +1,7 @@
 
 class PaymentTransaction {
   String tokenId;
-  double amount;
+  int amount;
 
   PaymentTransaction({this.tokenId, this.amount});
